Add time limit to bench.
PCx have a option to limite the duration of the search for the
solution. To have this option in the benchmark will need to
update the benchmark script to Python 3.3.1 or high and use the
subprocess class.

Some improves in the documentation of the benchmark script have
been done and the solution process of the current process file
are save in a file.