Correct "bug" in the return code.
For some problems Interior Points Methods have problems in the
convergence and fail.

For this problems the Solution->Status set up in the function PCx can
get news user in trouble because a conflict with the return code of
PCx program. In this commit we change the Solution->Status and the
return code of PCx program to be compatible.