Using C version of Ng-Peyton
We will be using the C version of Ng-Peyton and deprecating the
Fortran version.

Why? Aurelio Oliveira told that there is better using the C version.
For more info see
https://www.listas.unicamp.br/pipermail/pcx-dev-l/2013-October/000005.html.

The C version requires the F2C folder that already have been add in
a previous commit.

Other changes
-------------

* Update `configure.ac` file to handle the F2C folder.
* Update `Makefile.in`
* Replace the Old-Fashioned Suffix Rules
  (http://www.gnu.org/software/make/manual/html_node/Suffix-Rules.html#Suffix-Rules)
  with Pattern Rules
  (http://www.gnu.org/software/make/manual/html_node/Pattern-Rules.html#Pattern-Rules)
  in `Makefile.in`.

  Yes, there is some compatible problems using the Pattern Rules but I like new
  stuffs.
* Add `.PHONY` targets.
* When clean, using `\rm` instead of `rm` for safety.

  Yes, there is some compatible problems using `.PHONY` targes too.

DISCLAIMER
----------

This are a lot of changes in the build process and PCx source code.
I had successfully compile this version of PCx using

* autoconf 2.68
* make 3.81
* gcc 4.6.3

in a i686 machine running a GNU/Linux distribution and solving
`netlib-afiro`.

But when using

* autoconf 2.69
* make 4.0
* gcc 4.8.2

in a X86_64 machine (i3-3217U) it compile without problem but got a
Segmentation fault when solving `netlib-afiro`.

So I not guarantee that you will have success compiling PCx with
any tools and/or architecture.