Fix bug related with output directory
BUG DESCRIPTION:

The output will be write inside the `.OUT` directory. If this
directory not exist a segmentation fault happen.

If the output file can't be open with write permission a
segmentation fault happen because it try to write something
before check if the file was successfully open.

BUG SOLUTION:

* Add new option in the specification file to handle the output
  directory.
* Check if the output directory exist. If not create it.
* Check if the file was successfully open before try to write
  something.